+++ test/unit/tests/routes/addresses.unit.ts
@@ -28,6 +28,7 @@ describe('address service', () => {
 
       queryMock.onCall(0).resolves(fixture.sqlQueryMock);
       queryMock.onCall(1).resolves(fixture.sqlQueryMock2);
+      queryMock.onCall(2).resolves(fixture.sqlQueryMock3);
 
       const response = await supertest(fastify.server).get(fixture.endpoint);